move airflow integration files around
