+++ metadata-ingestion/src/datahub_provider/lineage/datahub.py
@@ -2,6 +2,7 @@ import json
 from typing import TYPE_CHECKING, Dict, List, Optional
 
 import dateutil.parser
+from airflow.configuration import conf
 from airflow.lineage.backend import LineageBackend
 
 import datahub.emitter.mce_builder as builder
@@ -9,17 +10,9 @@ import datahub.metadata.schema_classes as models
 
 if TYPE_CHECKING:
     from airflow import DAG
+    from airflow.models.baseoperator import BaseOperator
 
-    # from airflow.taskinstance import TaskInstance
-    from airflow.models.baseoperator import (
-        BaseOperator,
-    )  # pylint: disable=cyclic-import
-
-    from datahub.integrations.airflow.hooks import (
-        DatahubGenericHook,
-    )  # pylint: disable=cyclic-import
-
-from airflow.configuration import conf
+    from datahub.integrations.airflow.hooks import DatahubGenericHook
 
 
 def _entities_to_urn_list(iolets: List) -> List[str]:
@@ -51,7 +44,7 @@ class DatahubAirflowLineageBackend(LineageBackend):
             SerializedDAG,
         )
 
-        from datahub.integrations.airflow.hooks import AIRFLOW_1
+        from datahub_provider.hooks.datahub import AIRFLOW_1
 
         # Detect Airflow 1.10.x inlet/outlet configurations in Airflow 2.x, and
         # convert to the newer version. This code path will only be triggered
