+++ metadata-ingestion/tests/unit/test_airflow.py
@@ -73,12 +73,12 @@ def patch_airflow_connection(conn: Connection) -> Iterator[Connection]:
     # The return type should really by ContextManager, but mypy doesn't like that.
     # See https://stackoverflow.com/questions/49733699/python-type-hints-and-context-managers#comment106444758_58349659.
     with mock.patch(
-        "datahub.integrations.airflow.hooks.BaseHook.get_connection", return_value=conn
+        "datahub_provider.hooks.datahub.BaseHook.get_connection", return_value=conn
     ):
         yield conn
 
 
-@mock.patch("datahub.integrations.airflow.hooks.DatahubRestEmitter", autospec=True)
+@mock.patch("datahub_provider.hooks.datahub.DatahubRestEmitter", autospec=True)
 def test_datahub_rest_hook(mock_emitter):
     with patch_airflow_connection(datahub_rest_connection_config) as config:
         hook = DatahubRestHook(config.conn_id)
@@ -89,7 +89,7 @@ def test_datahub_rest_hook(mock_emitter):
         instance.emit_mce.assert_called_with(lineage_mce)
 
 
-@mock.patch("datahub.integrations.airflow.hooks.DatahubKafkaEmitter", autospec=True)
+@mock.patch("datahub_provider.hooks.datahub.DatahubKafkaEmitter", autospec=True)
 def test_datahub_kafka_hook(mock_emitter):
     with patch_airflow_connection(datahub_kafka_connection_config) as config:
         hook = DatahubKafkaHook(config.conn_id)
@@ -101,7 +101,7 @@ def test_datahub_kafka_hook(mock_emitter):
         instance.flush.assert_called_once()
 
 
-@mock.patch("datahub.integrations.airflow.operators.DatahubRestHook.emit_mces")
+@mock.patch("datahub.integrations.airflow.hooks.DatahubRestHook.emit_mces")
 def test_datahub_lineage_operator(mock_emit):
     with patch_airflow_connection(datahub_rest_connection_config) as config:
         task = DatahubEmitterOperator(
@@ -164,7 +164,7 @@ def test_hook_airflow_ui(hook):
     ],
 )
 @mock.patch("airflow.models.BaseOperator.xcom_push", autospec=True)
-@mock.patch("datahub.integrations.airflow.operators.DatahubRestHook.emit_mces")
+@mock.patch("datahub.integrations.airflow.hooks.DatahubRestHook.emit_mces")
 def test_lineage_backend(mock_emit, mock_xcom_push, inlets, outlets):
     DEFAULT_DATE = days_ago(2)
